Added basic holonomic drive implementation

+++ include/hardware/pros_motor.h
@@ -19,7 +19,7 @@ class ProsMotor : virtual public interface::Motor,
   void MoveAbsolute(double position, int max_velocity);
 
   void ResetEncoder();
-  int GetPosition();
+  double GetPosition();
   float GetVelocity();
 
  private:
