+++ include/interface/encoder.h
@@ -4,7 +4,7 @@ namespace interface {
 class Encoder {
  public:
   virtual void ResetEncoder();
-  virtual int GetPosition();
+  virtual double GetPosition();
   virtual float GetVelocity();
 };
 }  // namespace interface
